requirements.txt
paramiko==1.15.3
pyyaml==3.11